Support unsigned bigint auto-increment
Store generated BIGINT UNSIGNED AUTO_INCREMENT values above signed 64-bit range as exact decimal text at the SQLite boundary, preserve that metadata through catalog and ALTER paths, and cover information_schema and SHOW CREATE TABLE reporting.

+++ packages/libmylite/src/runtime/mylite_show_create_table_options.c
@@ -14,8 +14,12 @@ void mylite_show_create_table_append_options(
         collation == NULL ? mylite_charset_default_name() : collation->character_set;
 
     sqlite3_str_appendf(create_sql, " ENGINE=%s", info->engine == NULL ? "InnoDB" : info->engine);
-    if (info->has_auto_increment && info->auto_increment > 1) {
-        sqlite3_str_appendf(create_sql, " AUTO_INCREMENT=%lld", info->auto_increment);
+    if (info->has_auto_increment && info->auto_increment > 1U) {
+        sqlite3_str_appendf(
+            create_sql,
+            " AUTO_INCREMENT=%llu",
+            (unsigned long long)info->auto_increment
+        );
     }
     sqlite3_str_appendf(
         create_sql,
